o Go back to "make clean test" from "make depend test", looks like "make depend" produces some falce positives. We shall probably utilize CCACHE for speeding things up instead;
o re-run the test with every found superfluous include, since there
might be some inter-dependencies, so if I1 is tested befoe I2, and I2
is the only thing that depends on I1, by re-running we also catch
I1 as being superfluous. We would not catch it before;

o Fix output logic for multiple includes to be reported.